Rename collision_common and cmake organization
