+++ moveit_core/collision_detection_hpp_fcl/include/moveit/collision_detection_hpp_fcl/collision_common_hpp_fcl.h
